% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_data.R
\docType{data}
\name{dep_data}
\alias{dep_data}
\title{Representatives Data}
\format{An object of class \code{data.frame} with 4389 rows and 8 columns.}
\source{
\href{https://dadosabertos.camara.leg.br/swagger/api.html}{Dados Abertos Câmara}
}
\usage{
data(dep_data)
}
\description{
Data from Chamber of Representatives deputies. Used to query 
parties, states and deputy ID. Contains data from the 50th legislature onwards.
Aditional data can be queried from the official API.
}
\keyword{datasets}
