% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data}
\alias{get_data}
\title{Get Plenary Data}
\usage{
get_data(initial.date, final.date, format = "\%d/\%m/\%Y")
}
\arguments{
\item{initial.date}{(\code{character}) defining first date to include in the search}

\item{final.date}{(\code{character}) defining last date to include in the search}

\item{format}{specify which date format is being used as input}
}
\value{
a tibble with one row for each speech
}
\description{
Crawls trough plenary session transcripts between defined dates, downloads PDF files and parse data
}
\examples{
plenary_data <- get_data('20/03/2019', '27/03/2019', format = '\%d/\%m/\%Y')

}
